+++ bin/gitm-combine.js
@@ -2,9 +2,10 @@
 const program = require('commander')
 const sh = require('shelljs')
 const { options, args } = require('./conf/combine')
-const { error, queue, getStatus, getCurrent, searchBranchs, createArgs } = require('./js/index')
-const { defaults, appName } = require('./js/global')
-const config = require('./js/config')
+const { error, queue, getStatus, getCurrent, searchBranch, createArgs } = require('./js/index')
+const { defaults } = require('./js/global')
+const { appName } = require('./js/getGitConfig')()
+const config = require('./js/getConfig')()
 /**
  * gitm combine
  */