+++ bin/gitm-update.js
@@ -2,8 +2,8 @@
 const program = require('commander')
 const sh = require('shelljs')
 const { options, args } = require('./conf/update')
-const { error, queue, getStatus, getCurrent, searchBranchs, createArgs } = require('./js/index')
-const config = require('./js/config')
+const { error, queue, getStatus, getCurrent, searchBranch, createArgs } = require('./js/index')
+const config = require('./js/getConfig')()
 const { defaults } = require('./js/global')
 /**
  * gitm update