+++ bin/gitm.js
@@ -59,6 +59,10 @@ program
 	.command('clean', '提交权限')
 	.command('postmsg', '推送云之家消息')
 	.command('permission', '提交权限')
+	.command('hook', 'git钩子指令')
+	.command('run', 'git钩子运行指令')
+	.command('log', '查询日志')
+	.command('go', '智能猜测你要执行的动作')
 	.command('admin <command>', '管理员功能，包含对发版分支bugfix、release的操作')
 
 // 自定义帮助
@@ -71,7 +75,7 @@ program.on('--help', function () {
 
 // 映射不存在的指令
 program.on('command:*', function (types, opts) {
-	let cmd = ['init', 'config', 'combine', 'cb', 'start', 'st', 'end', 'ed', 'update', 'up', 'branch', 'bh', 'save', 'sv', 'get', 'gt', 'copy', 'cp', 'merge', 'mg', 'continue', 'ct', 'revert', 'rt', 'upgrade', 'ug', 'build', 'bd', 'ui', 'unlink', 'link', 'clean', 'postmsg', 'permission', 'admin']
+	let cmd = ['init', 'config', 'combine', 'cb', 'start', 'st', 'end', 'ed', 'update', 'up', 'branch', 'bh', 'save', 'sv', 'get', 'gt', 'copy', 'cp', 'merge', 'mg', 'continue', 'ct', 'revert', 'rt', 'upgrade', 'ug', 'build', 'bd', 'ui', 'unlink', 'link', 'clean', 'postmsg', 'permission', 'hook', 'run', 'log', 'go', 'admin']
 	if (!cmd.includes(types[0])) {
 		let arr = [].concat(types).concat(opts)
 		sh.exec('git ' + arr.join(' '), { silent: false })