+++ sh/build.js
@@ -1,1 +1,1 @@
-console.log(process.argv.slice(2));
+console.log(process.argv.slice(2))