+++ sh/commandSync.js
@@ -1,9 +1,0 @@
-const execa = require('execa');
-
-module.exports = (command, props) => {
-	try {
-		return execa.sync(command, props);
-	} catch (e) {
-		return e;
-	}
-};